+++ Documentation/template/public/main.js
@@ -1,0 +1,9 @@
+export default {
+    iconLinks: [
+        {
+          icon: 'github',
+          href: 'https://github.com/Odotocodot/Linq2OneNote',
+          title: 'GitHub'
+        }
+    ]
+}