+++ Odotocodot.OneNote.Linq/IOneNoteItem.cs
@@ -4,9 +4,12 @@ using System.Collections.Generic;
 namespace Odotocodot.OneNote.Linq
 {
     /// <summary>
-    /// The base interface of OneNote hierarchy items types. <br/>
-    /// (These are <see cref="OneNoteNotebook"/>, <see cref="OneNoteSectionGroup"/>, <see cref="OneNoteSection"/> and <see cref="OneNotePage"/>).
+    /// The base interface of OneNote hierarchy items types.
     /// </summary>
+    /// <seealso cref="OneNoteNotebook"/>
+    /// <seealso cref="OneNoteSectionGroup"/>
+    /// <seealso cref="OneNoteSection"/>
+    /// <seealso cref="OneNotePage"/>
     public interface IOneNoteItem
     {
         /// <summary>
@@ -26,8 +29,8 @@ namespace Odotocodot.OneNote.Linq
         /// </summary>
         DateTime LastModified { get; }
         /// <summary>
-        /// The direct children of the OneNote hierarchy <see cref="IOneNoteItem">item</see>, e.g. for a <see cref="OneNoteNotebook">notebook</see> it would be <see cref="OneNoteSection">sections</see> and/or <see cref="OneNoteSectionGroup">section groups</see>. <br/>
-        /// If the <see cref="IOneNoteItem">item</see> has no children an empty <see cref="IEnumerable{T}">IEnumerable</see>&lt;<see cref="IOneNoteItem"/>&gt; is returned.
+        /// The direct children of the OneNote hierarchy <see cref="IOneNoteItem">item</see>, e.g. for a <see cref="OneNoteNotebook">notebook</see> it could contain <see cref="OneNoteSection">sections</see> and/or <see cref="OneNoteSectionGroup">section groups</see>. <br/>
+        /// If the <see cref="IOneNoteItem">item</see> has no children an empty <see cref="IEnumerable{T}">IEnumerable</see>&lt;<see cref="IOneNoteItem"/>&gt; is returned. For instance, this property is an empty enumerable for a <see cref="OneNotePage">page</see>.
         /// </summary>
         IEnumerable<IOneNoteItem> Children { get; }
         /// <summary>