+++ Odotocodot.OneNote.Linq/Internal/IWritableHasIsInRecycleBin.cs
@@ -1,0 +1,9 @@
+using Odotocodot.OneNote.Linq.Abstractions;
+
+namespace Odotocodot.OneNote.Linq.Internal
+{
+	internal interface IWritableHasIsInRecycleBin : IHasIsInRecycleBin
+	{
+		new bool IsInRecycleBin { set; }
+	}
+}